+++ R/dummy.R
@@ -16,10 +16,8 @@
 #'  role should they be assigned?. By default, the function assumes
 #'  that the binary dummy variable columns created by the original
 #'  variables will be used as predictors in a model.
-#' @param contrast A specification for which type of contrast
-#'  should be used to make a set of full rank dummy variables. See
-#'  [stats::contrasts()] for more details. **not
-#'  currently working**
+#' @param one_hot A logical. For C levels, should C dummy variables be created
+#' rather than C-1?
 #' @param naming A function that defines the naming convention for
 #'  new dummy columns. See Details below.
 #' @param levels A list that contains the information needed to
@@ -58,6 +56,13 @@
 #'  Instead of values such as "`.L`", "`.Q`", or "`^4`", ordinal
 #'  dummy variables are given simple integer suffixes such as
 #'  "`_1`", "`_2`", etc.
+#'
+#' To change the type of contrast being used, change the global contrast option
+#'  via `options`.
+#'
+#' The [package vignette for dummy variables](https://topepo.github.io/recipes/articles/Dummies.html)
+#' and interactions has more information.
+#'
 #' @seealso [step_factor2string()], [step_string2factor()],
 #'  [dummy_names()], [step_regex()], [step_count()],
 #'  [step_ordinalscore()], [step_unorder()], [step_other()]
@@ -76,6 +81,20 @@
 #' unique(okc$diet)
 #' grep("^diet", names(dummy_data), value = TRUE)
 #'
+#' # Obtain the full set of dummy variables using `one_hot` option
+#' rec %>%
+#'   step_dummy(diet, one_hot = TRUE) %>%
+#'   prep(training = okc, retain = TRUE) %>%
+#'   juice(starts_with("diet")) %>%
+#'   names() %>%
+#'   length()
+#'
+#' length(unique(okc$diet))
+#'
+#' # Without one_hot
+#' length(grep("^diet", names(dummy_data), value = TRUE))
+#'
+#'
 #' tidy(dummies, number = 1)
 
 
@@ -84,7 +103,7 @@ step_dummy <-
            ...,
            role = "predictor",
            trained = FALSE,
-           contrast = options("contrasts"),
+           one_hot = FALSE,
            naming = dummy_names,
            levels = NULL,
            skip = FALSE) {
@@ -94,7 +113,7 @@ step_dummy <-
         terms = ellipse_check(...),
         role = role,
         trained = trained,
-        contrast = contrast,
+        one_hot = one_hot,
         naming = naming,
         levels = levels,
         skip = skip
@@ -106,7 +125,7 @@ step_dummy_new <-
   function(terms = NULL,
            role = "predictor",
            trained = FALSE,
-           contrast = contrast,
+           one_hot = one_hot,
            naming = naming,
            levels = levels,
            skip = FALSE
@@ -116,7 +135,7 @@ step_dummy_new <-
       terms = terms,
       role = role,
       trained = trained,
-      contrast = contrast,
+      one_hot = one_hot,
       naming = naming,
       levels = levels,
       skip = skip
@@ -144,7 +163,11 @@ prep.step_dummy <- function(x, training, info = NULL, ...) {
   levels <- vector(mode = "list", length = length(col_names))
   names(levels) <- col_names
   for (i in seq_along(col_names)) {
-    form <- as.formula(paste0("~", col_names[i]))
+    form_chr <- paste0("~", col_names[i])
+    if(x$one_hot) {
+      form_chr <- paste0(form_chr, "-1")
+    }
+    form <- as.formula(form_chr)
     terms <- model.frame(form,
                          data = training,
                          xlev = x$levels[[i]])
@@ -164,7 +187,7 @@ prep.step_dummy <- function(x, training, info = NULL, ...) {
     terms = x$terms,
     role = x$role,
     trained = TRUE,
-    contrast = x$contrast,
+    one_hot = x$one_hot,
     naming = x$naming,
     levels = levels,
     skip = x$skip
@@ -205,7 +228,10 @@ bake.step_dummy <- function(object, newdata, ...) {
     options(na.action = old_opt)
     on.exit(expr = NULL)
 
-    indicators <- indicators[, -1, drop = FALSE]
+    if(!object$one_hot) {
+      indicators <- indicators[, colnames(indicators) != "(Intercept)", drop = FALSE]
+    }
+
     ## use backticks for nonstandard factor levels here
     used_lvl <- gsub(paste0("^", col_names[i]), "", colnames(indicators))
     colnames(indicators) <- object$naming(col_names[i], used_lvl, fac_type == "ordered")