+++ R/recipe.R
@@ -172,8 +172,7 @@ recipe.data.frame <-
 
     if (!is_tibble(x))
       x <- as_tibble(x)
-    if (is.null(vars))
-      vars <- colnames(x)
+
     if (any(table(vars) > 1))
       stop("`vars` should have unique members", call. = FALSE)
     if (any(!(vars %in% colnames(x))))
@@ -335,8 +334,11 @@ prep.recipe <-
         stop("A training set must be supplied to the `training` argument ",
              "when `fresh = TRUE`", call. = FALSE)
       training <- x$template
-      tr_data <- train_info(training)
     } else {
+      if (!all(x$var_info$variable %in% colnames(training))) {
+        stop("Not all variables in the recipe are present in the supplied ",
+             "training set", call. = FALSE)
+      }
       training <- if (!is_tibble(training))
         as_tibble(training[, x$var_info$variable, drop = FALSE])
       else
@@ -425,12 +427,12 @@ bake <- function(object, ...)
 #'   returned by the function. See [selections()] for more details.
 #'   If no selectors are given, the default is to use
 #'   [everything()].
-#' @param composition Either "tibble", "matrix", or "dgCMatrix" for the
-#'  format of the processed data set. Note that all computations
-#'  during the baking process are done in a non-sparse format. Also,
-#'  note that this argument should be called **after** any selectors
-#'  and the selectors should only resolve to numeric columns
-#'  (otherwise an error is thrown).
+#' @param composition Either "tibble", "matrix", "data.frame", or
+#'  "dgCMatrix" for the format of the processed data set. Note that
+#'  all computations during the baking process are done in a
+#'  non-sparse format. Also, note that this argument should be
+#'  called **after** any selectors and the selectors should only
+#'  resolve to numeric columns (otherwise an error is thrown).
 #' @return A tibble, matrix, or sparse matrix that may have different
 #'  columns than the original columns in `newdata`.
 #' @details [bake()] takes a trained recipe and applies the
@@ -497,6 +499,8 @@ bake.recipe <- function(object, newdata, ..., composition = "tibble") {
     newdata <- convert_matrix(newdata, sparse = TRUE)
   } else if (composition == "matrix") {
     newdata <- convert_matrix(newdata, sparse = FALSE)
+  } else if (composition == "data.frame") {
+    newdata <- base::as.data.frame(newdata)
   }
 
   newdata
@@ -658,11 +662,13 @@ juice <- function(object, ..., composition = "tibble") {
     newdata <- convert_matrix(newdata, sparse = TRUE)
   } else if (composition == "matrix") {
     newdata <- convert_matrix(newdata, sparse = FALSE)
+  } else if (composition == "data.frame") {
+    newdata <- base::as.data.frame(newdata)
   }
 
   newdata
 }
 
-formats <- c("tibble", "dgCMatrix", "matrix")
+formats <- c("tibble", "dgCMatrix", "matrix", "data.frame")