+++ tests/testthat/test_dummies.R
@@ -3,7 +3,6 @@ library(recipes)
 
 data(okc)
 
-okc$location <- gsub(", california", "", okc$location)
 okc$diet[is.na(okc$diet)] <- "missing"
 okc <- okc[complete.cases(okc), -5]
 
@@ -12,7 +11,7 @@ okc_fac$diet <- factor(okc_fac$diet)
 okc_fac$location <- factor(okc_fac$location)
 
 test_that('dummy variables with factor inputs', {
-  rec <- recipe(age ~ ., data = okc_fac)
+  rec <- recipe(age ~ location + diet, data = okc_fac)
   dummy <- rec %>% step_dummy(diet, location)
   dummy_trained <- prep(dummy, training = okc_fac, verbose = FALSE, stringsAsFactors = FALSE)
   dummy_pred <- bake(dummy_trained, newdata = okc_fac, all_predictors())
@@ -20,7 +19,7 @@ test_that('dummy variables with factor inputs', {
   dummy_pred <- as.data.frame(dummy_pred)
   rownames(dummy_pred) <- NULL
 
-  exp_res <- model.matrix(age ~ ., data = okc_fac)[, -1]
+  exp_res <- model.matrix(age ~ location + diet, data = okc_fac)[, -1]
   exp_res <- exp_res[, colnames(exp_res) != "age"]
   colnames(exp_res) <- gsub("^location", "location_", colnames(exp_res))
   colnames(exp_res) <- gsub("^diet", "diet_", colnames(exp_res))
@@ -38,11 +37,32 @@ test_that('dummy variables with factor inputs', {
 })
 
 test_that('dummy variables with string inputs', {
-  rec <- recipe(age ~ ., data = okc)
+  rec <- recipe(age ~ location + diet, data = okc)
   dummy <- rec %>% step_dummy(diet, location)
   expect_error(
     prep(dummy, training = okc, verbose = FALSE, stringsAsFactors = FALSE)
   )
+})
+
+test_that('create all dummy variables', {
+  rec <- recipe(age ~ location + diet + height, data = okc_fac)
+  dummy <- rec %>% step_dummy(diet, location, one_hot = TRUE)
+  dummy_trained <- prep(dummy, training = okc_fac, verbose = FALSE, stringsAsFactors = FALSE)
+  dummy_pred <- bake(dummy_trained, newdata = okc_fac, all_predictors())
+  dummy_pred <- dummy_pred[, order(colnames(dummy_pred))]
+  dummy_pred <- as.data.frame(dummy_pred)
+  rownames(dummy_pred) <- NULL
+
+  exp_res <- NULL
+  for(pred in c("diet", "height", "location")) {
+    tmp <- model.matrix(as.formula(paste("~", pred, "+ 0")), data = okc_fac)
+    colnames(tmp) <- gsub(paste0("^", pred), paste0(pred, "_"), colnames(tmp))
+    exp_res <- bind_cols(exp_res, as_tibble(tmp))
+  }
+  colnames(exp_res) <- make.names(colnames(exp_res))
+  exp_res <- as.data.frame(exp_res)
+  rownames(exp_res) <- NULL
+  expect_equivalent(dummy_pred, exp_res)
 })
 
 test_that('tests for issue #91', {
@@ -69,8 +89,8 @@ test_that('tests for issue #91', {
 
 test_that('naming function', {
   expect_equal(dummy_names("x", letters[1:3]), c("x_a", "x_b", "x_c"))
-  expect_equal(dummy_names("x", letters[1:3], ordinal = TRUE), 
-               c("x_1", "x_2", "x_3")) 
+  expect_equal(dummy_names("x", letters[1:3], ordinal = TRUE),
+               c("x_1", "x_2", "x_3"))
 })
 
 test_that('printing', {